backend/requirements.txt
fastapi==0.116.1
gunicorn==23.0.0
openai==1.99.9
pydantic==2.11.7
pyyaml==6.0.2
uvicorn==0.35.0
